Migrate innerpanel
